Create Temp
This project is about  a class AssassinManager that allows a client to manage a game of assassin.  Each person playing assassin has a particular target that he/she is trying to assassinate.  One of the things that makes the game more interesting to play is that initially, each person knows only who they are assassinating (they don't know who is trying to assassinate them nor do they know who other people are trying to assassinate).  You are working on a program for the “game administrator” who needs to keep track of who is stalking whom and the history of who killed whom.